Final implementation for interactive movie website
My team's website has pages for the following: 
a "home" page with a description of our purpose,
a "bookmark" page for creating an exportable list of movies to watch, a review page for viewing/ writing movie reviews, 
an "external website" page to view professional review resources (ex: Rotten Tomatoes film reviews), 
and finally, a "popular" page to view five random movies from "IMDB Top 100 movies of all time".
This was a fun project and taught me about good practices in web development. I was not as focused on the appearance of the website as I was on experimenting with various concepts (APIs, Local Storage, etc.). Hope you enjoy it!